#6549 - Apache Tika can not parse Microsoft Docx format in native mode - move reflections maven artifact under the tika-deployment module
