+++ extensions/tika/deployment/pom.xml
@@ -12,6 +12,10 @@
     <artifactId>quarkus-tika-deployment</artifactId>
     <name>Quarkus - Apache Tika - Deployment</name>
 
+    <properties>
+        <reflections.version>0.9.12</reflections.version>
+    </properties>
+
     <dependencies>
         <dependency>
             <groupId>io.quarkus</groupId>
@@ -25,10 +29,13 @@
             <groupId>io.quarkus</groupId>
             <artifactId>quarkus-tika</artifactId>
         </dependency>
+        <!-- reflections -->
         <dependency>
             <groupId>org.reflections</groupId>
             <artifactId>reflections</artifactId>
+            <version>${reflections.version}</version>
         </dependency>
+
         <dependency>
             <groupId>xalan</groupId>
             <artifactId>xalan</artifactId>
