+++ extensions/tika/runtime/src/main/java/io/quarkus/tika/runtime/TikaConfiguration.java
@@ -1,11 +1,11 @@
 package io.quarkus.tika.runtime;
 
+import java.util.Map;
+import java.util.Optional;
+
 import io.quarkus.runtime.annotations.ConfigItem;
 import io.quarkus.runtime.annotations.ConfigPhase;
 import io.quarkus.runtime.annotations.ConfigRoot;
-
-import java.util.Map;
-import java.util.Optional;
 
 /**
  * Tika parser configuration
